+++ example/lib/main.dart
@@ -1,7 +1,14 @@
+import 'dart:io';
+
 import 'package:work/work.dart';
+import 'package:json_annotation/json_annotation.dart';
+
+import 'simple_work.dart';
+
+part 'main.g.dart';
 
 void main() async {
-  final data = await TestWork().start(['xxx']);
+  final data = await TestWork('xxx').start();
 
   if (data.success) {
     print(data.result);
@@ -10,28 +17,37 @@ void main() async {
     print(data.errorCode);
   }
 
-  final download = await DownloadWork().start([
-    'file:/xxx/test.jpg',
-    'key',
-    123,
-  ]);
+  final download = await DownloadWork(
+    path: 'file:/xxx/test.jpg',
+    key: 'key',
+    resNo: 123,
+  ).start();
 
   if (download.success) {
     // show('file:/xxx/test.jpg')
   }
+
+  final upload = await UploadWork(File('test.jpg')).start();
+
+  if (upload.success) {
+    print(upload.result);
+  }
 }
 
 class TestWork extends SimpleWork<String> {
+  TestWork(this.param1);
+
+  /// 请求参数1
+  final String param1;
+
   @override
   String onExtractResult(resultData, data) => resultData['account'];
 
   @override
-  String onUrl(List params) => 'https://api.example.com/test';
+  String onUrl() => 'https://api.example.com/test';
 
   @override
-  void onFillParams(Map<String, dynamic> data, List params) {
-    data['param1'] = params[0];
-  }
+  Map<String, dynamic> onFillParams() => {'param1': param1};
 
   @override
   String onNetworkError(data) => '网络连接失败，当前网络不可用';
@@ -43,22 +59,52 @@ class TestWork extends SimpleWork<String> {
   String onParseFailed(data) => '请求失败，服务器异常';
 
   @override
-  String onRequestFailedMessage(data) => data.response.data['message'] ?? '操作失败';
+  String onRequestFailedMessage(data) =>
+      data.response!.data['message'] ?? '操作失败';
 
   @override
-  String onRequestSuccessMessage(data) => data.response.data['message'] ?? '';
+  String onRequestSuccessMessage(data) => data.response!.data['message'] ?? '';
 }
 
+@JsonSerializable()
 class DownloadWork extends SimpleDownloadWork {
-  @override
-  void onFillParams(Map<String, dynamic> data, List params) {
-    data['key'] = params[1];
-    data['resNo'] = params[2];
-  }
+  DownloadWork({required this.path, required this.key, required this.resNo});
+
+  /// 存放路径
+  @JsonKey(ignore: true)
+  final String path;
+
+  /// 请求参数key
+  final String key;
+
+  /// 请求参数resNo
+  final int resNo;
 
   @override
-  String onDownloadPath(List params) => params[0];
+  Map<String, dynamic> onFillParams() => _$DownloadWorkToJson(this);
 
   @override
-  String onUrl(List params) => 'https://api.example.com/test.jpg';
+  String onDownloadPath() => path;
+
+  @override
+  String onUrl() => 'https://api.example.com/test.jpg';
+}
+
+@JsonSerializable()
+class UploadWork extends SimpleWork<String> {
+  UploadWork(this.file);
+
+  /// 需要上传的文件
+  @JsonKey(toJson: workFileToJsonConvert)
+  final File file;
+
+  /// 假设返回的结果"result"标签对应的是文件的网络地址
+  @override
+  String onExtractResult(resultData, data) => resultData;
+
+  @override
+  String onUrl() => 'https://api.example.com/upload';
+
+  @override
+  Map<String, dynamic> onFillParams() => _$UploadWorkToJson(this);
 }