+++ example/lib/main.g.dart
@@ -1,0 +1,18 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'main.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+Map<String, dynamic> _$DownloadWorkToJson(DownloadWork instance) =>
+    <String, dynamic>{
+      'key': instance.key,
+      'resNo': instance.resNo,
+    };
+
+Map<String, dynamic> _$UploadWorkToJson(UploadWork instance) =>
+    <String, dynamic>{
+      'file': workFileToJsonConvert(instance.file),
+    };