+++ example/lib/simple_work.dart
@@ -3,11 +3,7 @@
 import 'dart:async';
 
 import 'package:meta/meta.dart';
-import 'work_model.dart';
-import 'work_core.dart';
-
-/// 用于获取响应json数据协议中"result"字段
-const String result = 'result';
+import 'package:work/work.dart';
 
 /// 简化的[WorkData]类实现
 ///
@@ -28,7 +24,7 @@ const String result = 'result';
 /// ```
 class SimpleWorkData<T> extends WorkData<T> {
   /// 协议错误码
-  int _errorCode;
+  int _errorCode = 0;
 
   /// 协议错误码
   int get errorCode => _errorCode;
@@ -54,59 +50,57 @@ class SimpleWorkData<T> extends WorkData<T> {
 ///
 /// ```
 abstract class SimpleWork<D> extends Work<D, SimpleWorkData<D>> {
+  /// 用于获取响应json数据协议中"result"字段
+  static const String result = 'result';
+
   @override
   SimpleWorkData<D> onCreateWorkData() => SimpleWorkData<D>();
 
   @override
-  FutureOr<bool> onCheckResponse(SimpleWorkData<D> data) =>
-      data.response.data != null;
-
-  @override
-  FutureOr<D> onResponseSuccess(SimpleWorkData<D> data) {
-    if (data.response.data[result] == null) {
+  FutureOr<D?> onRequestSuccess(SimpleWorkData<D> data) {
+    if (data.response?.data?[result] == null) {
       return onDefaultResult(data);
     } else {
-      return onExtractResult(data.response.data[result], data);
+      return onExtractResult(data.response!.data[result], data);
     }
   }
 
   @override
-  FutureOr<bool> onResponseResult(SimpleWorkData<D> data) =>
-      data.response.data['state'];
+  bool onRequestResult(SimpleWorkData<D> data) => data.response!.data['state'];
 
   @mustCallSuper
   @override
-  FutureOr<D> onRequestFailed(SimpleWorkData<D> data) {
-    if (data.response.data['errorCode'] != null) {
-      data._errorCode = data.response.data['errorCode'];
+  FutureOr<D?> onRequestFailed(SimpleWorkData<D> data) {
+    if (data.response!.data['errorCode'] != null) {
+      data._errorCode = data.response!.data['errorCode']!;
     }
 
     return super.onRequestFailed(data);
   }
 
   @override
-  FutureOr<String> onRequestSuccessMessage(SimpleWorkData<D> data) =>
-      data.response.data['message'];
+  String? onRequestSuccessMessage(SimpleWorkData<D> data) =>
+      data.response!.data['message'];
 
   @override
-  FutureOr<String> onRequestFailedMessage(SimpleWorkData<D> data) =>
-      data.response.data['message'];
+  String? onRequestFailedMessage(SimpleWorkData<D> data) =>
+      data.response!.data['message'];
 
   /// 生成响应成功的结果数据
   ///
   /// * [resultData]为协议中的[result]标签下的数据
   /// * 当请求成功且返回结果中存在[result]标签且不为null时被调用
   /// * 返回装配后的本地数据对象
-  /// * [data]为将要返回的数据包装类，包含有传入的参数[data.params]
+  /// * [data]为将要返回的数据包装类
   @protected
-  FutureOr<D> onExtractResult(resultData, SimpleWorkData<D> data);
+  FutureOr<D?> onExtractResult(resultData, SimpleWorkData<D> data);
 
   /// 生成响应成功的默认结果数据
   ///
   /// * 当请求成功且返回结果不存在[result]标签或值为null时被调用，默认实现为null
-  /// * [data]为将要返回的数据包装类，包含有传入的参数[data.params]
+  /// * [data]为将要返回的数据包装类
   @protected
-  FutureOr<D> onDefaultResult(SimpleWorkData<D> data) => null;
+  FutureOr<D?> onDefaultResult(SimpleWorkData<D> data) => null;
 }
 
 /// 简化的下载专用[Work]类
@@ -117,22 +111,20 @@ abstract class SimpleDownloadWork extends Work<void, SimpleWorkData<void>> {
   SimpleWorkData<void> onCreateWorkData() => SimpleWorkData<void>();
 
   @override
-  FutureOr<void> onResponseSuccess(SimpleWorkData<void> data) => null;
+  bool onRequestResult(SimpleWorkData<void> data) => true;
 
   @override
-  FutureOr<bool> onResponseResult(SimpleWorkData<void> data) => true;
+  FutureOr<void> onRequestSuccess(SimpleWorkData<void> data) => null;
 
   @override
-  HttpMethod get httpMethod => HttpMethod.download;
+  HttpMethod onHttpMethod() => HttpMethod.download;
 
   @mustCallSuper
   @override
-  FutureOr<void> onConfigOptions(Options options, List params) {
-    options.downloadPath = onDownloadPath(params);
+  FutureOr<void> onConfigOptions(Options options) {
+    options.downloadPath = onDownloadPath();
   }
 
-  /// 设置下载文件路径
-  ///
-  /// [params]为任务传入参数，返回下载文件要保存的位置路径
-  FutureOr<String> onDownloadPath(List params);
+  /// 返回下载文件路径
+  String onDownloadPath();
 }