+++ lib/src/_convert_native.dart
@@ -4,13 +4,14 @@ import 'dart:io';
 
 import 'package:dio/dio.dart';
 import 'package:http_parser/http_parser.dart';
+
 import 'work_model.dart';
 
 /// 用于[com.HttpMethod.upload]请求类型的数据转换
 ///
 /// [src]原始参数，返回处理后的符合dio接口的参数
 Future<FormData> convertToDio(Map<String, dynamic> src) async {
-  Future<dynamic> onConvert(value) async {
+  Future<dynamic> onConvert(dynamic value) async {
     if (value is File) {
       value = UploadFileInfo(value.path);
     }
@@ -18,17 +19,19 @@ Future<FormData> convertToDio(Map<String, dynamic> src) async {
     if (value is UploadFileInfo) {
       if (value.stream != null) {
         return MultipartFile(
-          value.stream,
-          value.length,
+          value.stream!,
+          value.length!,
           filename: value.fileName,
-          contentType: MediaType.parse(value.mimeType),
+          contentType:
+              value.mimeType != null ? MediaType.parse(value.mimeType!) : null,
         );
       }
 
       return MultipartFile.fromFile(
-        value.filePath,
+        value.filePath!,
         filename: value.fileName,
-        contentType: MediaType.parse(value.mimeType),
+        contentType:
+            value.mimeType != null ? MediaType.parse(value.mimeType!) : null,
       );
     }